finally got the the bad_line and line_num of the skipped lines from the warnings
====================

> ./convert_to_csai.py
cs8.txt done ....on to the next one....
cs9.txt done ....on to the next one....
cs1.txt done ....on to the next one....
bad_line is # 14 ... A1|Are there any upcoming events for the [Stat-Club] club in the STAT department?|(No. | Yes. The next event is [Stat-Event] on [Date].)

cs2.txt done ....on to the next one....
cs3.txt done ....on to the next one....
cs7.txt done ....on to the next one....
cs6.txt done ....on to the next one....
bad_line is # 14 ... A1|Are there any upcoming events for the [Stat-Club] club in the STAT department?|(No. | Yes. The next event is [Stat-Event] on [Date].)

cs4.txt done ....on to the next one....
cs5.txt done ....on to the next one....
cs15.txt done ....on to the next one....
cs14.txt done ....on to the next one....
cs10.txt done ....on to the next one....
cs11.txt done ....on to the next one....
cs13.txt done ....on to the next one....
bad_line is # 31 ... G1|How many units of tech electives are required for a [Major| Minor]?|Number

cs12.txt done ....on to the next one....
stat1.txt done ....on to the next one....
clubs.txt done ....on to the next one....
> cat STACIA_QA/cs2.txt | head -n 16
A1|What clubs does the [Department] department offer?|The following clubs are in the [Department] department: [Club], �.
A1|How many clubs are in the [Department] department?|There are [Number] clubs.
A1|Does [Faculty] advise any clubs in the [Department] department?|[Faculty] is the advisor of the [Club] club.
A1|Who is the advisor of the [Club] club?|[Faculty] is the advisor of the [Club] club.
A1|What is the email of the President of the [Club] club?|[Person]�s email is [Email].
A1|What type of club is the [Club] club?|It is a [Type] club.
A1|What types of clubs does the [Department] department offer?|[Club(s)] are (is a) [Type] club(s),�.
A1|What is the email for the advisor of the [Club] club?|[Faculty]�s email is [Email].
A1|What is the [Club] club?|[Club-Description]
A1|What is the homepage of the [Club] club?|The homepage is [Club-Website].
A1|What is the box number of the [Club] club?|The box number to the [Club] club is [Box-Number].
A1|What is the email for the [Club] club?|The email for the [Club] club is [Email].
A1|What are the social media accounts for the [Club] club?|[Club] club�s Facebook page is [Facebook]. [Club] club�s Instagram account is [Instagram].
A1|Are there any upcoming events for the [Stat-Club] club in the STAT department?|(No. | Yes. The next event is [Stat-Event] on [Date].)
A1|How many people follow the [Stat-Club] club�s Facebook page?|[Followers] follow the [Stat-Club] club�s Facebook page.
A1|Are there any new posts on the [Stat-Club] club�s Facebook page?|There are [Posts] new posts today and [Posts] posts in the past 30 days on the [Stat-Club] club�s Facebook page.
>